+++ app/Http/Controllers/Admin/PermissionController.php
@@ -3,12 +3,16 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Traits\AuthorizationsTrait;
 use App\Models\Permission;
 use App\Http\Requests\StorePermissionRequest;
 use App\Http\Requests\UpdatePermissionRequest;
+use Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
+    use AuthorizationsTrait;
     /**
      * Display a listing of the resource.
      *
@@ -16,9 +20,12 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        Gate::authorize('managePermissions', Permission::class);
+
         $data['active'] = 'permissions';
         $data['title'] = 'اجازه ها';
-        $data['permissions'] = Permission::paginate(10);
+        $data['permissions'] = Permission::orderBy("ability", "asc")->paginate(10);
+        $data['manuAuthorizations'] = $this->getMenuAuthorizations(Auth::user());
 
         return view('admin.permission.all', $data);
     }
@@ -41,9 +48,13 @@ class PermissionController extends Controller
      */
     public function store(StorePermissionRequest $request)
     {
+        Gate::authorize('create', Permission::class);
+
         $permission = new Permission;
 
         $permission->name = $request->name;
+        $permission->ability = $request->ability;
+        $permission->description = $request->description;
 
         $permission->save();
 
@@ -74,6 +85,7 @@ class PermissionController extends Controller
         $data['active'] = 'permissions';
         $data['title'] = 'ویرایش اجازه ' . $permission->name;
         $data['permissions'] = Permission::paginate(10);
+        $data['manuAuthorizations'] = $this->getMenuAuthorizations(Auth::user());
 
         return view('admin.permission.all', $data);
     }
@@ -87,7 +99,11 @@ class PermissionController extends Controller
      */
     public function update(UpdatePermissionRequest $request, Permission $permission)
     {
+        Gate::authorize('update', $permission);
+
         $permission->name = $request->name;
+        $permission->ability = $request->ability;
+        $permission->description = $request->description;
         $permission->save();
 
         return redirect('permissions')->with('success_message', 'اجازه با موفقیت ویرایش شد.');
@@ -101,6 +117,8 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        Gate::authorize('delete', $permission);
+
         $permission->roles()->detach();
 
         $permission->delete();
